rename files to nodejs standards
- files should be lowercase
- added post install grunt exec
