+++ src/library/factory/weather-factory-interface.ts
@@ -1,4 +1,4 @@
-import {WeatherInterface} from "./../models/weatherInterface";
+import {WeatherInterface} from "./../models/weather-interface";
 
 export interface WeatherFactoryInterface {
     create(): WeatherInterface;
