+++ src/library/factory/weather-factory.ts
@@ -1,5 +1,5 @@
-import {WeatherFactoryInterface} from "./weatherFactoryInterface";
-import {WeatherInterface} from "./../models/weatherInterface";
+import {WeatherFactoryInterface} from "./weather-factory-interface";
+import {WeatherInterface} from "./../models/weather-interface";
 import {Weather} from "./../models/weather";
 
 export class WeatherFactory implements WeatherFactoryInterface {
