+++ src/library/mapper/json-to-weather-mapper-interface.ts
@@ -1,4 +1,4 @@
-import {WeatherInterface} from "./../models/WeatherInterface";
+import {WeatherInterface} from "./../models/weather-interface";
 
 export interface JsonToWeatherMapperInterface {
     covert(data: string): WeatherInterface;
