+++ src/library/models/weather-interface.ts
