+++ src/library/services/http-connection/http-request-interface.ts
