+++ src/library/services/weather/weather-service-interface.ts
@@ -1,4 +1,4 @@
-import {WeatherInterface} from "./../../models/weatherInterface";
+import {WeatherInterface} from "./../../models/weather-interface";
 
 export interface WeatherServiceInterface {
   getTemperatureInDegreeCelsiusByLatitudeAndLongitude(lat: number, long: number): Promise<WeatherInterface>;
